# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Factory Inventory Management System","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/inventory":{"get":{"summary":"Get Inventory","description":"Get all inventory items with optional filtering","operationId":"get_inventory_api_inventory_get","parameters":[{"name":"warehouse","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Warehouse"}},{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InventoryItem"},"title":"Response Get Inventory Api Inventory Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inventory/{item_id}":{"get":{"summary":"Get Inventory Item","description":"Get a specific inventory item","operationId":"get_inventory_item_api_inventory__item_id__get","parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"string","title":"Item Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventoryItem"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/orders":{"get":{"summary":"Get Orders","description":"Get all orders with optional filtering","operationId":"get_orders_api_orders_get","parameters":[{"name":"warehouse","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Warehouse"}},{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"month","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Month"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Order"},"title":"Response Get Orders Api Orders Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/orders/{order_id}":{"get":{"summary":"Get Order","description":"Get a specific order","operationId":"get_order_api_orders__order_id__get","parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"string","title":"Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/demand":{"get":{"summary":"Get Demand Forecasts","description":"Get demand forecasts","operationId":"get_demand_forecasts_api_demand_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DemandForecast"},"type":"array","title":"Response Get Demand Forecasts Api Demand Get"}}}}}}},"/api/backlog":{"get":{"summary":"Get Backlog","description":"Get backlog items with purchase order status","operationId":"get_backlog_api_backlog_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/BacklogItem"},"type":"array","title":"Response Get Backlog Api Backlog Get"}}}}}}},"/api/dashboard/summary":{"get":{"summary":"Get Dashboard Summary","description":"Get summary statistics for dashboard with optional filtering","operationId":"get_dashboard_summary_api_dashboard_summary_get","parameters":[{"name":"warehouse","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Warehouse"}},{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"month","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Month"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/spending/summary":{"get":{"summary":"Get Spending Summary","description":"Get spending summary statistics","operationId":"get_spending_summary_api_spending_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/spending/monthly":{"get":{"summary":"Get Monthly Spending","description":"Get monthly spending breakdown","operationId":"get_monthly_spending_api_spending_monthly_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/spending/categories":{"get":{"summary":"Get Category Spending","description":"Get spending by category","operationId":"get_category_spending_api_spending_categories_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/spending/transactions":{"get":{"summary":"Get Recent Transactions","description":"Get recent transactions","operationId":"get_recent_transactions_api_spending_transactions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/reports/quarterly":{"get":{"summary":"Get Quarterly Reports","description":"Get quarterly performance reports","operationId":"get_quarterly_reports_api_reports_quarterly_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/reports/monthly-trends":{"get":{"summary":"Get Monthly Trends","description":"Get month-over-month trends","operationId":"get_monthly_trends_api_reports_monthly_trends_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"BacklogItem":{"properties":{"id":{"type":"string","title":"Id"},"order_id":{"type":"string","title":"Order Id"},"item_sku":{"type":"string","title":"Item Sku"},"item_name":{"type":"string","title":"Item Name"},"quantity_needed":{"type":"integer","title":"Quantity Needed"},"quantity_available":{"type":"integer","title":"Quantity Available"},"days_delayed":{"type":"integer","title":"Days Delayed"},"priority":{"type":"string","title":"Priority"},"has_purchase_order":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Has Purchase Order","default":false}},"type":"object","required":["id","order_id","item_sku","item_name","quantity_needed","quantity_available","days_delayed","priority"],"title":"BacklogItem"},"DemandForecast":{"properties":{"id":{"type":"string","title":"Id"},"item_sku":{"type":"string","title":"Item Sku"},"item_name":{"type":"string","title":"Item Name"},"current_demand":{"type":"integer","title":"Current Demand"},"forecasted_demand":{"type":"integer","title":"Forecasted Demand"},"trend":{"type":"string","title":"Trend"},"period":{"type":"string","title":"Period"}},"type":"object","required":["id","item_sku","item_name","current_demand","forecasted_demand","trend","period"],"title":"DemandForecast"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InventoryItem":{"properties":{"id":{"type":"string","title":"Id"},"sku":{"type":"string","title":"Sku"},"name":{"type":"string","title":"Name"},"category":{"type":"string","title":"Category"},"warehouse":{"type":"string","title":"Warehouse"},"quantity_on_hand":{"type":"integer","title":"Quantity On Hand"},"reorder_point":{"type":"integer","title":"Reorder Point"},"unit_cost":{"type":"number","title":"Unit Cost"},"location":{"type":"string","title":"Location"},"last_updated":{"type":"string","title":"Last Updated"}},"type":"object","required":["id","sku","name","category","warehouse","quantity_on_hand","reorder_point","unit_cost","location","last_updated"],"title":"InventoryItem"},"Order":{"properties":{"id":{"type":"string","title":"Id"},"order_number":{"type":"string","title":"Order Number"},"customer":{"type":"string","title":"Customer"},"items":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Items"},"status":{"type":"string","title":"Status"},"order_date":{"type":"string","title":"Order Date"},"expected_delivery":{"type":"string","title":"Expected Delivery"},"total_value":{"type":"number","title":"Total Value"},"actual_delivery":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Actual Delivery"},"warehouse":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Warehouse"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"}},"type":"object","required":["id","order_number","customer","items","status","order_date","expected_delivery","total_value"],"title":"Order"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}